requirements.txt
livekit-agents[silero,turn-detector,groq]~=1.2
livekit-plugins-noise-cancellation~=0.2
livekit-plugins-elevenlabs
python-dotenv
sounddevice
